# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App-Api",
        "version": "1.0.0"
    },
    "paths": {
        "/api/columns": {
            "get": {
                "tags": [
                    "Columns"
                ],
                "summary": "Get all Columns",
                "operationId": "88b58fc2f509f8e8f7bf9729230efa9c",
                "responses": {
                    "200": {
                        "description": "A list of all projects"
                    }
                }
            },
            "post": {
                "tags": [
                    "Columns"
                ],
                "summary": "Create a new project",
                "operationId": "fee53761a4d8f3c3f737d1c61af32508",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "projectId",
                                    "order"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Column Name"
                                    },
                                    "projectId": {
                                        "description": "The ID of the user associated with the project",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order": {
                                        "description": "The URL or file path of the project's background image",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The created project"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/columns/{id}": {
            "get": {
                "tags": [
                    "Columns"
                ],
                "summary": "Get a column by ID",
                "operationId": "297201ea06f6001f129b6ba21bd8a26f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the column to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested column",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Column 1"
                                        },
                                        "projectId": {
                                            "type": "integer",
                                            "example": 101
                                        },
                                        "order": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Column not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Columns"
                ],
                "summary": "Update a column by ID",
                "operationId": "a2a8844ce9281406552f8b596a2b06c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the column to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "projectId",
                                    "order"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Updated name of the column",
                                        "type": "string",
                                        "example": "Updated Column"
                                    },
                                    "projectId": {
                                        "description": "Updated project ID",
                                        "type": "integer",
                                        "example": 102
                                    },
                                    "order": {
                                        "description": "Updated order of the column",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Column updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Updated Column"
                                        },
                                        "projectId": {
                                            "type": "integer",
                                            "example": 102
                                        },
                                        "order": {
                                            "type": "integer",
                                            "example": 2
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Column not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Columns"
                ],
                "summary": "Delete a column by ID",
                "operationId": "99a8529dddf9f1c747fb9443dde027fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the column to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Column deleted successfully"
                    },
                    "404": {
                        "description": "Column not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}